partie voyages simple

+++ src/com/projet/model/Airport.java
@@ -2,11 +2,12 @@ package com.projet.model;
 
 public class Airport {
 	
-	private long id;
-	 public long getId() {
+	private int id;
+	
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	private String nom;
